Refactor llms
