+++ README.md
@@ -30,12 +30,12 @@ To install development dependencies, run:
 
 **Set your API information in llm.yaml**
 
-First, create an LLM config template by running the `debug-gym-init-llm-config` entrypoint:
+First, create an LLM config template by running `python -m debug_gym.llms.configure`:
 
-    python -m debug_gym.init_llm_config ~/.config/debug_gym
+    python -m debug_gym.llms.configure
 
 > [!TIP]
-> Run `debug-gym-init-llm-config --help` for more options. By default, the template is created at `~/.config/debug_gym/llm.yaml`, but you can specify any directory.
+> Run `python -m debug_gym.llms.configure --help` for more options. By default, the template is created at `$HOME/.config/debug_gym/llm.yaml`, but you can specify any directory.
 
 Then, edit this file with your endpoint and credentials. You can choose one of these authentication methods:
 - For authenticating with an API key, provide `api_key`.
@@ -44,7 +44,7 @@ Then, edit this file with your endpoint and credentials. You can choose one of t
 > [!WARNING]
 > When using open-sourced LLMs, e.g., via vLLM, you need to correctly setup `HF_TOKEN` required by the tokenizer.
 
-By default, `debug-gym` looks for the LLM config file at `~/.config/debug_gym/llm.yaml`. You can change this behavior by exporting the environment variable `LLM_CONFIG_FILE_PATH` or by setting `llm_config_file_path` in your script config file (see [Running Baselines](#3-running-baselines)).
+By default, `debug-gym` looks for the LLM config file at `$HOME/.config/debug_gym/llm.yaml`. You can change this behavior by exporting the environment variable `LLM_CONFIG_FILE_PATH` or by setting `llm_config_file_path` in your script config file (see [Running Baselines](#3-running-baselines)).
 
 ---
 
@@ -57,12 +57,15 @@ debug_gym
 │   ├── envs
 │   ├── terminal
 │   └── tools
-└── agents
+├── agents
+└── llms
 ```
 
 `debug_gym.gym` is a simulation environment. Given a code repository, an agent can iteratively interact with a set of tools, such as `pdb`, that are designed for investigate the code. Once gathered enough information, the agent can propose a patch that rewrites certain lines of the code. The terminal will subsequently execute the new code against a set of test cases.
 
 `debug_gym.agents` are LLM-based debugging agents that use `debug_gym.gym` to interact with code repositories to seek necessary information and thus fix potential bugs. At an interaction step, the agent takes a text observation that describes the environment states and tool states as input, it is expected to generate a command, subsequently, the environment will provide a new text observation in response, describing the state change caused by that command.
+
+`debug_gym.llms` are the different LLM backends that can be used to instantiate agents. Currently, we support OpenAI, Azure OpenAI, and Anthropic.
 
 ---
 
@@ -143,8 +146,8 @@ As an example, we provide a buggy pytorch code repository in `data/pytorch`.
 #### 3.5. Analysis and Visualization
 
 We provide a set of scripts to help analyze the log files (e.g., the `.jsonl` files) generated by the agent.
-- In the `analysis` folder, we provide scripts that used to generate the corresponding figures in our technical report. 
-- In the `analysis/json_log_viewer` folder, we provide a Flask app to view a `.jsonl` log file in the browser. 
+- In the `analysis` folder, we provide scripts that used to generate the corresponding figures in our technical report.
+- In the `analysis/json_log_viewer` folder, we provide a Flask app to view a `.jsonl` log file in the browser.
 
 ## Citation
 ```
