+++ debug_gym/agents/base_agent.py
@@ -8,7 +8,7 @@ from os.path import join as pjoin
 import numpy as np
 
 from debug_gym.agents.history_tracker import HistoryTracker, build_history_prompt
-from debug_gym.agents.llm_api import LLM
+from debug_gym.llms.base import LLM
 from debug_gym.agents.utils import trim
 from debug_gym.gym.envs.env import RepoEnv
 from debug_gym.gym.utils import unescape
