+++ debug_gym/agents/history_tracker.py
@@ -1,7 +1,7 @@
 import copy
 from dataclasses import asdict
 
-from debug_gym.agents.llm_api import LLM, LLMResponse
+from debug_gym.llms.base import LLM, LLMResponse
 from debug_gym.gym.envs.env import EnvInfo
