+++ debug_gym/llms/configure.py
@@ -2,7 +2,8 @@ import argparse
 import os
 from pathlib import Path
 
-from debug_gym.agents.llm_api import LLM_CONFIG_TEMPLATE
+from termcolor import colored
+from debug_gym.llms.constants import LLM_CONFIG_TEMPLATE, DEFAULT_LLM_CONFIG
 
 
 def init_llm_config(dest_dir: str = None):
@@ -10,10 +11,11 @@ def init_llm_config(dest_dir: str = None):
     directory or the user's home directory."""
 
     parser = argparse.ArgumentParser(
-        description="Create an LLM config template in the specified directory or `~/.config/debug_gym`."
+        description="Create an LLM config template."
     )
     parser.add_argument(
-        "destination", nargs="?", type=str, help="Destination directory (positional)"
+        "destination", nargs="?", type=str,
+        help=f"Destination directory (positional). Defaults to {DEFAULT_LLM_CONFIG.parent}"
     )
     parser.add_argument("-d", "--dest", type=str, help="Destination directory")
     parser.add_argument(
@@ -47,7 +49,13 @@ def init_llm_config(dest_dir: str = None):
     else:
         print(f"LLM config template already exists at `{destination}`.")
 
-    print("Please edit the file to configure your LLM settings.")
+    print(
+        colored(
+            f"Please edit `{destination}` to configure your LLM settings.",
+            "green",
+            attrs=["bold"],
+        )
+    )
 
 
 if __name__ == "__main__":
