+++ tests/agents/test_example_agent.py
@@ -1,7 +1,7 @@
 from unittest.mock import MagicMock
 
 from debug_gym.agents.debug_agent import Debug_5_Agent, DebugAgent
-from debug_gym.agents.llm_api import LLMResponse, TokenUsage
+from debug_gym.llms.base import LLMResponse, TokenUsage
 from debug_gym.agents.rewrite_agent import RewriteAgent
