+++ tests/llms/test_utils.py
@@ -1,0 +1,12 @@
+
+from debug_gym.llms.utils import print_messages
+
+
+def test_print_messages(logger_mock):
+    messages = [
+        {"role": "user", "content": "Hello"},
+        {"role": "assistant", "content": "Hi"},
+        {"role": "system", "content": "System message"},
+    ]
+    print_messages(messages, logger_mock)
+    assert logger_mock._log_history == ["Hello\n", "Hi\n", "System message\n"]
